requirements.txt
pandas
confluent-kafka
pyspark
pytest
pytest-mock
python-dotenv